+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.2.1
+- Fixed issue with uptime sensor showing second of uptime.
+- Added timestamp and uptime seconds attributes to uptime sensor.
+- NAS's with muliple NICs now show the mac addresses for each NIC not just the first one.
+  
 ## v1.2.0
 
 - Added Model, Serial number and Firmware version to the device.